Changes to read/write signatures required to build ed and ecp. (#751)
Also more consistent with related functions.

+++ sys/src/libc/9sys/write.c
@@ -10,8 +10,8 @@
 #include <u.h>
 #include <libc.h>
 
-ssize_t
-write(int d, const void *buf, size_t nbytes)
+int32_t
+write(int d, const void *buf, int32_t nbytes)
 {
 	int n;
